2022.9.27-1

+++ src/Client.java
@@ -4,7 +4,7 @@ public class Client {
     public static void main(String[] args) {
         Scanner input = new Scanner(System.in);
         System.out.print("Enter number to be parsed: ");
-        int num = input.nextInt();
+        long num = input.nextLong();
         System.out.println(Functions.modChallengeLoop(num));
     }
 }
